tools/ack-discover/requirements.txt
boto3==1.21.8
GitPython==3.1.26
PyYAML==5.4
PyGitHub==1.55
prettytable==3.2.0
